[Bronze V] Title: 이상한 기호, Time: 168 ms, Memory: 17644 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -1,16 +1,16 @@
 import java.util.Scanner;
 
 public class Main {
-    static int calculate(int a, int b) {
-        int ans = (a + b) * (a - b);
+    static long calculate(long a, long b) {
+        long ans = (a + b) * (a - b);
         return ans;
     }
 
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
 
-        int a = sc.nextInt();
-        int b = sc.nextInt();
+        long a = sc.nextLong();
+        long b = sc.nextLong();
 
         System.out.println(calculate(a, b));
     }
